+++ C-prototipos/Pruebas_Unitarias/RTC-DS1302/.vscode/settings.json
@@ -1,0 +1,3 @@
+{
+    "C_Cpp.errorSquiggles": "disabled"
+}
