+++ C-prototipos/Pruebas_Unitarias/RTC-DS1302/desktop.ini
@@ -1,0 +1,2 @@
+[.ShellClassInfo]
+LocalizedResourceName=RTC-DS1302
